Move common functions to the CHTMLFontData class (so they can be used in both tex and other fonts without duplication)
