+++ mathjax3-ts/output/svg/fonts/tex.ts
@@ -22,7 +22,7 @@
  */
 
 import {SVGFontData, SVGFontDataClass, SVGCharOptions, SVGVariantData, DelimiterData,
-        CharData, CharOptions, DelimiterMap, CharMapMap, FontDataClass} from '../FontData.js';
+        CharData, CharOptions, DelimiterMap, CharMapMap, CssFontMap, FontDataClass} from '../FontData.js';
 import {CommonTeXFontMixin} from '../../common/fonts/tex.js';
 import {StyleList, StyleData} from '../../common/CssStyles.js';
 import {em} from '../../../util/lengths.js';
@@ -62,6 +62,22 @@ import {delimiters} from '../../common/fonts/tex/delimiters.js';
  *  The TeXFont class
  */
 export class TeXFont extends CommonTeXFontMixin<SVGCharOptions, SVGVariantData, SVGFontDataClass>(SVGFontData) {
+
+    /**
+     * The data used for CSS for undefined characters each variant
+     */
+    protected static defaultCssFonts: CssFontMap = {
+        ...SVGFontData.defaultCssFonts,
+        '-smallop': ['serif', false, false],
+        '-largeop': ['serif', false, false],
+        '-size3': ['serif', false, false],
+        '-size4': ['serif', false, false],
+        '-tex-caligraphic': ['cursive', true, false],
+        '-tex-bold-caligraphic': ['cursive', true, true],
+        '-tex-oldstyle': ['serif', false, false],
+        '-tex-bold-oldstyle': ['serif', false, true],
+        '-tex-mathit': ['serif', true, false]
+    };
 
     /**
      *  The stretchy delimiter data
